refactor: Update shader code to use uint for cameraSID and improve type consistency
- Modified shader functions to utilize `uint` for `cameraSID`, enhancing type safety and consistency across shader implementations.
- Adjusted related shader logic to ensure compatibility with the new type requirements, improving overall code clarity and maintainability.

+++ samples/test_e2e/CustomShader/main.ts
@@ -65,11 +65,11 @@ function createCustomShader() {
 
   /* shaderity: @{mainPrerequisites} */
 
-    mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
-    mat4 viewMatrix = get_viewMatrix(cameraSID);
-    mat4 projectionMatrix = get_projectionMatrix(cameraSID);
-    mat3 normalMatrix = get_normalMatrix(a_instanceInfo.x);
-    bool isBillboard = get_isBillboard(a_instanceInfo.x);
+    mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
+    mat4 viewMatrix = get_viewMatrix(uint(cameraSID));
+    mat4 projectionMatrix = get_projectionMatrix(uint(cameraSID));
+    mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
+    bool isBillboard = get_isBillboard(uint(a_instanceInfo.x));
 
     // Skeletal
     processGeometry(
@@ -93,7 +93,7 @@ function createCustomShader() {
     v_texcoord_0 = a_texcoord_0;
     v_baryCentricCoord = a_baryCentricCoord.xyz;
 
-    bool visibility = get_isVisible(a_instanceInfo.x);
+    bool visibility = get_isVisible(uint(a_instanceInfo.x));
     if (!visibility)
     {
       gl_Position = vec4(0.0);
@@ -138,7 +138,7 @@ function createCustomShader() {
     // Normal
     vec3 normal_inWorld = normalize(v_normal_inWorld);
 
-    vec4 diffuseColorFactor = get_diffuseColorFactor(materialSID, 0);
+    vec4 diffuseColorFactor = get_diffuseColorFactor(uint(materialSID), 0u);
 
 
     // diffuseColor (Considered to be premultiplied alpha)
@@ -157,8 +157,8 @@ function createCustomShader() {
     }
 
     // diffuseColorTexture (Considered to be premultiplied alpha)
-    vec4 diffuseColorTextureTransform = get_diffuseColorTextureTransform(materialSID, 0);
-    float diffuseColorTextureRotation = get_diffuseColorTextureRotation(materialSID, 0);
+    vec4 diffuseColorTextureTransform = get_diffuseColorTextureTransform(uint(materialSID), 0u);
+    float diffuseColorTextureRotation = get_diffuseColorTextureRotation(uint(materialSID), 0u);
     vec2 diffuseColorTexUv = uvTransform(diffuseColorTextureTransform.xy, diffuseColorTextureTransform.zw, diffuseColorTextureRotation, v_texcoord_0);
     vec4 textureColor = texture(u_diffuseColorTexture, diffuseColorTexUv);
     diffuseColor *= textureColor.rgb;
@@ -169,7 +169,7 @@ function createCustomShader() {
     // Lighting
     vec3 shadingColor = vec3(0.0, 0.0, 0.0);
   #ifdef RN_IS_LIGHTING
-    int shadingModel = get_shadingModel(materialSID, 0);
+    int shadingModel = get_shadingModel(uint(materialSID), 0u);
     if (shadingModel > 0) {
 
       vec3 diffuse = vec3(0.0, 0.0, 0.0);
@@ -185,10 +185,10 @@ function createCustomShader() {
         // Diffuse
         diffuse += diffuseColor * max(0.0, dot(normal_inWorld, light.direction)) * light.attenuatedIntensity;
 
-        float shininess = get_shininess(materialSID, 0);
-        int shadingModel = get_shadingModel(materialSID, 0);
+        float shininess = get_shininess(uint(materialSID), 0u);
+        int shadingModel = get_shadingModel(uint(materialSID), 0u);
 
-        vec3 viewPosition = get_viewPosition(cameraSID);
+        vec3 viewPosition = get_viewPosition(uint(cameraSID));
 
         // Specular
         if (shadingModel == 2) {// BLINN
